refactor(frontend): extract shared skeleton page header
Both page skeletons repeated the PageLayout header, and the card-grid and
dashboard skeletons need the same thing. Pull it into
SkeletonPageHeader so breadcrumb, title and FAB geometry lives in one
place instead of four.

Breadcrumbs are opt-in via breadcrumbCount, matching PageLayout, which
only introduces the 20px stack gap when a trail is present. Each skeleton
keeps its own placeholder text widths.
